GEMS-225: Implemented TSolModMulti class to initialize and manage Phase models.
