+++ tsolmod4rkt-build/Thermo-time-in/series1-dat.lst
@@ -1,0 +1,1 @@
+-f "series1-dch.json" "series1-ipm.json"  "series1.FUN.json"  "series1-dbr-0-0000.json"
