+++ tsolmod4rkt-build/Thermo-time-in/series1-dbr-0-0000.json
@@ -1,0 +1,15 @@
+[
+{
+  "_key" : "series1-dbr-0-0000",
+  "set" : "series1",
+  "dbr" : {
+    "NodeStatusFMT": -1,
+    "NodeStatusCH": 1,
+    "TK": 773.15,
+    "P": 100000000,
+    "bIC": [
+        0.29442571790996, 1.78324960377229e-08, 0.01, 1.11026667785195, 0.0872223855045395, 0.178347345796736, 3.15066041476453, 1.01307661056411, 0
+    ]
+  }
+}
+]
