+++ tsolmod4rkt-build/Thermo-time-in/series1-ipm.json
@@ -1,0 +1,49 @@
+[
+{
+  "_key" : "series1-ipm",
+  "set" : "series1",
+  "ipm" : {
+    "ID_key": "Solvus    G  series1     0001 0       500     400     0",
+    "pa_DK": 4e-05,
+    "pa_IIM": 7000,
+    "pa_PRD": -5,
+    "pa_PLLG": 13000,
+    "sMod": [
+        "HNNSNNNN", "VNNSNNNN", "VNNSNNNN"
+    ],
+    "LsMod": [
+        2, 0, 4, 
+        3, 2, 3, 
+        3, 2, 3
+    ],
+    "IPxPH": [
+        0, 1, 0, 2, 1, 2, 0, 1, 0, 2, 
+        1, 2
+    ],
+    "PMc": [
+        0.0299999993294477, 3.72000002861023, 1, 1, 0, 0, 0, 0, 3100, 0, 
+        0, 25100, -10.8000001907349, 0.342999994754791, 40000, 0, 0, 3100, 0, 0, 
+        25100, -10.8000001907349, 0.342999994754791, 40000, 0, 0
+    ],
+    "LsMdc": [
+        0, 0, 0, 
+        1, 0, 0, 
+        1, 0, 0
+    ],
+    "DMc": [
+        0.64300000667572, 1, 1, 0.64300000667572, 1, 1
+    ],
+    "B": [
+        0.29442571790996, 1.78324960377229e-08, 0.01, 1.11026667785195, 0.0872223855045395, 0.178347345796736, 3.15066041476453, 1.01307661056411, 0
+    ],
+    "dcMod": [
+        "HKF", "HKF", "HKF", "HKF", "HKF", "HKF", "HKF", "HKF", "HKF", "HKF", 
+        "HKF", "HKF", "HKF", "HKF", "HKF", "HKF", "HKF", "HKF", "HKF", "HKF", 
+        "HKF", "HKF", "HKF", "HKF", "HKF", "HKF", "HKF", "HKF", "HKF", "HKF", 
+        "HKF", "HKF", "HKF", "HKF", "HKF", "HWF", "CHE", "CHE", "CSE", "CHE", 
+        "CHE", "CSE", "CSE", "CSE", "CSE", "CSE", "CSE", "CSE", "CSE", "CSE", 
+        "CSE", "CHE", "CHE", "CHE", "CHE", "CSE", "CSE", "CSE", "CSE", "CSE"
+    ]
+  }
+}
+]
